HW2

+++ Lesson2/L2_task_3.py
@@ -1,4 +1,4 @@
-side = int(input("сторона="))
+side = float(input("сторона="))
 
 import math
 def square(side):
